normalized xlsx converter as intermediate class
