+++ common/base_converter/__init__.py
@@ -6,6 +6,7 @@ Use of this source code is governed by an MIT-style license that can be found in
 from .base_converter import BaseConverter
 from .csv_converter import CsvConverter
 from .json_converter import JsonConverter
+from .normalized_xlsx_converter import NormalizedXlsxConverter
 from .pull_converter import PullConverter
 from .xlsx_converter import XlsxConverter
 from .xml_converter import XmlConverter
